Moved blocks to wagtailspace app.
